+++ apps/api/src/services/configuration/configuration.service.ts
@@ -12,6 +12,8 @@ export class ConfigurationService {
     this.environmentConfiguration = cleanEnv(process.env, {
       ACCESS_TOKEN_SALT: str(),
       ALPHA_VANTAGE_API_KEY: str({ default: '' }),
+      API_KEY_COINGECKO_DEMO: str({ default: '' }),
+      API_KEY_COINGECKO_PRO: str({ default: '' }),
       BETTER_UPTIME_API_KEY: str({ default: '' }),
       CACHE_QUOTES_TTL: num({ default: 1 }),
       CACHE_TTL: num({ default: 1 }),